TASK-58 Refactor Defence Module
Разбиение логики defence модуля, перенос логики защиты в модуль core, переменование
модуля defence в модуль meeting, инкапсуляция логики связанной с созданием встреч
