--- defence/src/main/kotlin/ru/surf/defence/configuration/KafkaConsumerPropertiesConfiguration.kt
@@ -1,12 +1,0 @@
-package ru.surf.defence.configuration
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties(prefix = "spring.kafka.consumer")
-class KafkaConsumerPropertiesConfiguration(
-    val clientId: String,
-    val bootstrapServers: String,
-    val groupId: String,
-    val autoOffsetReset: String
-) {
-}
