--- defence/src/main/kotlin/ru/surf/defence/configuration/KafkaProducerPropertiesConfiguration.kt
@@ -1,9 +1,0 @@
-package ru.surf.defence.configuration
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties(prefix = "spring.kafka.producer")
-data class KafkaProducerPropertiesConfiguration(
-    val clientId: String,
-    val bootstrapServers: String
-)
