--- defence/src/main/kotlin/ru/surf/defence/configuration/ZoomAdminPropertiesConfiguration.kt
@@ -1,12 +1,0 @@
-package ru.surf.defence.configuration
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties(prefix = "zoom.admin")
-data class ZoomAdminPropertiesConfiguration(
-    val adminId: String,
-    val adminReserveId: String
-) {
-    var authorizationBasicToken: String = ""
-    var authorizationBearerToken: String = ""
-}
