--- defence/src/main/kotlin/ru/surf/defence/configuration/ZoomPropertiesConfiguration.kt
@@ -1,13 +1,0 @@
-package ru.surf.defence.configuration
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties(prefix = "zoom")
-data class ZoomPropertiesConfiguration(
-    val apiUrl: String,
-    val baseUrl: String,
-    val accountId: String,
-    val clientId: String,
-    val clientSecret: String
-) {
-}
