--- defence/src/main/kotlin/ru/surf/defence/dto/FullResponseDefenceDto.kt
@@ -1,3 +1,0 @@
-package ru.surf.defence.dto
-
-data class FullResponseDefenceDto(val some:String)
