--- defence/src/main/kotlin/ru/surf/defence/dto/ShortResponseDefenceDto.kt
@@ -1,4 +1,0 @@
-package ru.surf.defence.dto
-
-data class ShortResponseDefenceDto(val some:String) {
-}
