--- defence/src/main/kotlin/ru/surf/defence/repository/DefenceRepository.kt
@@ -1,11 +1,0 @@
-package ru.surf.defence.repository
-
-import org.springframework.data.jpa.repository.JpaRepository
-import ru.surf.defence.entity.Defence
-import java.util.*
-
-interface DefenceRepository : JpaRepository<Defence, UUID> {
-
-    fun findDefenceByZoomMeetingId(zoomMeetingId: Long): Defence?
-
-}
