--- defence/src/main/kotlin/ru/surf/defence/service/DefenceService.kt
@@ -1,13 +1,0 @@
-package ru.surf.defence.service
-
-import ru.surf.core.kafkaEvents.defence.CreateDefenceEvent
-import ru.surf.defence.entity.Defence
-
-interface DefenceService {
-
-    fun saveDefence(createDefenceEvent: CreateDefenceEvent): Defence
-
-    fun deleteDefenceByMeetingId(zoomMeetingId: Long)
-
-    fun getDefenceByMeetingId(zoomMeetingId: Long): Defence
-}
