--- defence/src/main/kotlin/ru/surf/defence/service/KafkaService.kt
@@ -1,7 +1,0 @@
-package ru.surf.defence.service
-
-interface KafkaService {
-
-    fun sendCoreEvent(event: Any)
-
-}
