--- defence/src/main/kotlin/ru/surf/defence/service/StrategyService.kt
@@ -1,7 +1,0 @@
-package ru.surf.defence.service
-
-interface StrategyService {
-
-    fun consumeEvent(event: Any): Any
-
-}
