--- defence/src/main/resources/application-prod.yml
@@ -1,41 +1,0 @@
-server:
-  port: ${PORT:8080}
-  servlet:
-    context-path: /core
-
-spring:
-  datasource:
-    url: jdbc:postgresql://${DB_HOST}/${DB_NAME}
-    username: ${DB_USER}
-    password: ${DB_PASSWORD}
-    driver-class-name: org.postgresql.Driver
-  jpa:
-    database-platform: org.hibernate.dialect.PostgreSQLDialect
-    hibernate:
-      ddl-auto: update
-    show-sql: true
-    properties:
-      hibernate.format_sql: true
-  flyway:
-    enabled: true
-    locations: db/migration/core
-  kafka:
-    producer:
-      bootstrap-servers: ${KAFKA_HOST}
-      key-serializer: org.apache.kafka.common.serialization.StringSerializer
-      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
-      acks: all
-      client-id: service-core
-      properties:
-        spring.json.add.type.headers: false
-
-springdoc:
-  swagger-ui:
-    enabled: true
-
-authService:
-  credentialsServiceApi: ${AUTH_API}
-
-externalFiles:
-  s3FileServiceApi: ${EXT_FILES_API}
-  s3ResourceServiceApi: ${EXT_FILES_RES_API}
