+++ meeting/src/test/kotlin/ru/surf/meeting/MeetingApplicationTests.kt
@@ -1,10 +1,10 @@
-package ru.surf.defence
+package ru.surf.meeting
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest
-class DefenceApplicationTests {
+class MeetingApplicationTests {
 
     @Test
     fun contextLoads() {
